geo-rep: Fix status transition
The status transition is as below which is
wrong.

Created->Initializing->Active->Active/Passive->Stopped

As soon as the monitor spawns the worker, the state
is changed from 'Initializing' to 'Active' and then to
'Active/Passive' based on whether worker gets the lock
or not. This is wrong and it should directly tranistion
as below.

Created->Initializing->Active/Passive->Stopped

> Change-Id: Ibf5ca5c4fdf168c403c6da01db60b93f0604aae7
> BUG: 1500284
> Signed-off-by: Kotresh HR <khiremat@redhat.com>
(cherry picked from commit 3edf926a1bda43879c09694cf3904c214c94c9dc)

Change-Id: Ibf5ca5c4fdf168c403c6da01db60b93f0604aae7
BUG: 1500835
Signed-off-by: Kotresh HR <khiremat@redhat.com>